mybatis study 1

+++ simple-Mybatis-1/src/main/java/com/tk/simple/mybatis/entity/CountryEntity.java
@@ -1,19 +1,17 @@
 package com.tk.simple.mybatis.entity;
 
-import lombok.Data;
-
 public class CountryEntity {
-    private Integer id;
+    private Long id;
 
     private String countryCode;
 
     private String countryName;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
